+++ scripts/Test-StateForgePackageMetadata.ps1
@@ -37,7 +37,8 @@ try {
         'src\StateForge.CloudNative\StateForge.CloudNative.csproj',
         'src\StateForge.Format\StateForge.Format.csproj',
         'src\StateForge.Prometheus\StateForge.Prometheus.csproj',
-        'src\StateForge.Performance\StateForge.Performance.csproj'
+        'src\StateForge.Performance\StateForge.Performance.csproj',
+        'src\StateForge.Replication\StateForge.Replication.csproj'
     )
 
     $errors = New-Object System.Collections.Generic.List[string]